Moves buttons to separate files
