+++ src/lib/Scenes/Artwork/Components/CommercialButtons.tsx
@@ -1,11 +1,10 @@
 import { Button, Spacer } from "@artsy/palette"
 import { CommercialButtons_artwork } from "__generated__/CommercialButtons_artwork.graphql"
-import { CommercialButtonsOfferOrderMutation } from "__generated__/CommercialButtonsOfferOrderMutation.graphql"
-import { CommercialButtonsOrderMutation } from "__generated__/CommercialButtonsOrderMutation.graphql"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import React from "react"
-import { Alert } from "react-native"
-import { commitMutation, createFragmentContainer, graphql, RelayProp } from "react-relay"
+import { createFragmentContainer, graphql, RelayProp } from "react-relay"
+import { BuyNowButtonFragmentContainer as BuyNowButton } from "./BuyNowButton"
+import { MakeOfferButtonFragmentContainer as MakeOfferButton } from "./MakeOfferButton"
 
 export interface CommercialButtonProps {
   artwork: CommercialButtons_artwork
@@ -13,207 +12,27 @@ export interface CommercialButtonProps {
   editionSetID?: string
 }
 
-export interface State {
-  isCommittingCreateOrderMutation: boolean
-  isCommittingCreateOfferOrderMutation: boolean
-}
-
-export class CommercialButtons extends React.Component<CommercialButtonProps, State> {
-  state = {
-    isCommittingCreateOfferOrderMutation: false,
-    isCommittingCreateOrderMutation: false,
-  }
-
-  onMutationError(orderType: "offer" | "order", error) {
-    Alert.alert("Sorry, we couldn't process the request.", "Please try again or contact orders@artsy.net for help.", [
-      {
-        text: "Cancel",
-        style: "cancel",
-      },
-      {
-        text: "Retry",
-        onPress: () => {
-          if (orderType === "order") {
-            this.handleCreateOrder()
-          } else if (orderType === "offer") {
-            this.handleCreateOfferOrder()
-          }
-        },
-      },
-    ])
-    console.log("src/lib/Scenes/Artwork/Components/CommercialButtons.tsx", error)
-  }
-
-  handleCreateOrder() {
-    const { relay, artwork, editionSetID } = this.props
-    const { isCommittingCreateOrderMutation, isCommittingCreateOfferOrderMutation } = this.state
-    const { internalID } = artwork
-    if (isCommittingCreateOrderMutation || isCommittingCreateOfferOrderMutation) {
-      return
-    }
-    this.setState({ isCommittingCreateOrderMutation: true }, () => {
-      if (relay && relay.environment) {
-        commitMutation<CommercialButtonsOrderMutation>(relay.environment, {
-          mutation: graphql`
-            mutation CommercialButtonsOrderMutation($input: CommerceCreateOrderWithArtworkInput!) {
-              commerceCreateOrderWithArtwork(input: $input) {
-                orderOrError {
-                  __typename
-                  ... on CommerceOrderWithMutationSuccess {
-                    order {
-                      internalID
-                      mode
-                    }
-                  }
-                  ... on CommerceOrderWithMutationFailure {
-                    error {
-                      type
-                      code
-                      data
-                    }
-                  }
-                }
-              }
-            }
-          `,
-          variables: {
-            input: {
-              artworkId: internalID,
-              editionSetId: editionSetID,
-            },
-          },
-          onCompleted: data => {
-            this.setState({ isCommittingCreateOrderMutation: false }, () => {
-              const {
-                commerceCreateOrderWithArtwork: { orderOrError },
-              } = data
-              if (orderOrError.error) {
-                this.onMutationError("order", orderOrError.error)
-              } else {
-                SwitchBoard.presentModalViewController(this, `/orders/${orderOrError.order.internalID}`)
-              }
-            })
-          },
-          onError: error =>
-            this.setState({ isCommittingCreateOrderMutation: false }, () => this.onMutationError("order", error)),
-        })
-      }
-    })
-  }
-
-  handleCreateOfferOrder() {
-    const { relay, artwork, editionSetID } = this.props
-    const { isCommittingCreateOfferOrderMutation, isCommittingCreateOrderMutation } = this.state
-    const { internalID } = artwork
-    if (isCommittingCreateOfferOrderMutation || isCommittingCreateOrderMutation) {
-      return
-    }
-    this.setState({ isCommittingCreateOfferOrderMutation: true }, () => {
-      if (relay && relay.environment) {
-        commitMutation<CommercialButtonsOfferOrderMutation>(relay.environment, {
-          mutation: graphql`
-            mutation CommercialButtonsOfferOrderMutation($input: CommerceCreateOfferOrderWithArtworkInput!) {
-              commerceCreateOfferOrderWithArtwork(input: $input) {
-                orderOrError {
-                  __typename
-                  ... on CommerceOrderWithMutationSuccess {
-                    order {
-                      internalID
-                      mode
-                    }
-                  }
-                  ... on CommerceOrderWithMutationFailure {
-                    error {
-                      type
-                      code
-                      data
-                    }
-                  }
-                }
-              }
-            }
-          `,
-          variables: {
-            input: {
-              artworkId: internalID,
-              editionSetId: editionSetID,
-            },
-          },
-          onCompleted: data => {
-            this.setState({ isCommittingCreateOfferOrderMutation: false }, () => {
-              const {
-                commerceCreateOfferOrderWithArtwork: { orderOrError },
-              } = data
-              if (orderOrError.error) {
-                this.onMutationError("offer", orderOrError.error)
-              } else {
-                SwitchBoard.presentModalViewController(this, `/orders/${orderOrError.order.internalID}`)
-              }
-            })
-          },
-          onError: error =>
-            this.setState({ isCommittingCreateOfferOrderMutation: false }, () => this.onMutationError("order", error)),
-        })
-      }
-    })
-  }
-
+export class CommercialButtons extends React.Component<CommercialButtonProps> {
   handleInquiry = () => {
     SwitchBoard.presentModalViewController(this, `/inquiry/${this.props.artwork.slug}`)
   }
 
   renderButtons = () => {
-    const { isAcquireable, isOfferable, isInquireable } = this.props.artwork
-    const { isCommittingCreateOrderMutation, isCommittingCreateOfferOrderMutation } = this.state
+    const { artwork } = this.props
+    const { isAcquireable, isOfferable, isInquireable } = artwork
+
     if (isOfferable && isAcquireable) {
       return (
         <>
-          <Button
-            onPress={() => this.handleCreateOrder()}
-            loading={isCommittingCreateOrderMutation}
-            size="large"
-            block
-            width={100}
-          >
-            Buy now
-          </Button>
+          <BuyNowButton artwork={artwork} editionSetID={this.props.editionSetID} />
           <Spacer mb={1} />
-          <Button
-            onPress={() => this.handleCreateOfferOrder()}
-            loading={isCommittingCreateOfferOrderMutation}
-            variant="secondaryOutline"
-            size="large"
-            block
-            width={100}
-          >
-            Make offer
-          </Button>
+          <MakeOfferButton artwork={artwork} editionSetID={this.props.editionSetID} variant="secondaryOutline" />
         </>
       )
     } else if (isAcquireable) {
-      return (
-        <Button
-          onPress={() => this.handleCreateOrder()}
-          loading={isCommittingCreateOrderMutation}
-          size="large"
-          block
-          width={100}
-        >
-          Buy now
-        </Button>
-      )
+      return <BuyNowButton artwork={artwork} editionSetID={this.props.editionSetID} />
     } else if (isOfferable) {
-      return (
-        <Button
-          onPress={() => this.handleCreateOfferOrder()}
-          loading={isCommittingCreateOfferOrderMutation}
-          size="large"
-          block
-          width={100}
-        >
-          Make offer
-        </Button>
-      )
+      return <MakeOfferButton artwork={artwork} editionSetID={this.props.editionSetID} />
     } else if (isInquireable) {
       return (
         <Button onPress={() => this.handleInquiry()} size="large" block width={100}>
@@ -238,6 +57,9 @@ export const CommercialButtonsFragmentContainer = createFragmentContainer(Commer
       isAcquireable
       isOfferable
       isInquireable
+
+      ...BuyNowButton_artwork
+      ...MakeOfferButton_artwork
     }
   `,
 })
